+++ popkorn/src/main/kotlin/cc/popkorn/Constants.kt
@@ -1,6 +1,8 @@
 package cc.popkorn
 
 
+val ALTERNATE_JAVA_LANG_PACKAGE = "java.lang" to "cc.popkorn.lang"
+
 const val RESOLVER_SUFFIX = "Resolver"
 const val PROVIDER_SUFFIX = "Provider"